Fix battery status for old laptops
There are 3 battery capacity values, curent, max, and design. Current is the
current capacity of the battery, design is the capacity that the battery was
designed to hold, and max is the maximum capacity that the battery can store.

The new logic uses the minimum of max and design as the divisor in the battery
percentage calculation.